added blur when logging in added more CSS skinning started work on the view for the employee scheduling column added a query to get appointments for an employee for their specific schedule column
Database is using Time but the appointment object is using Timestamp
Which one should we use?

+++ src/MiscObjects/Queries.java
@@ -112,7 +112,7 @@ public class Queries {
         String firstName;
         String lastName;
         String email;
-        double phoneNumber;
+        long phoneNumber;
         Date dateOfBirth;
 
         try {
@@ -128,7 +128,7 @@ public class Queries {
                 username1 = resultSet.getString("USERNAME");
                 firstName = resultSet.getString("F_NAME");
                 lastName = resultSet.getString("L_NAME");
-                phoneNumber = resultSet.getDouble("PHONE");
+                phoneNumber = resultSet.getLong("PHONE");
                 email = resultSet.getString("EMAIL");
                 dateOfBirth = resultSet.getDate("DATE_OF_BIRTH");
                 Employee employee = new Employee();
